prime number

+++ src/HelloCpp/headers/Pmath.h
@@ -1,3 +1,3 @@
 #pragma once
 
-bool IsPrime(const unsigned long number);
+bool IsPrime(const unsigned long long number);
